requirements_auto.txt
pandas
requests
python-dotenv

